// src/pages/GestionTecnologia/types.ts
export interface Componente {
  title: string;
  focus: string;
  critical: string;
  icon: string;
}

export interface TalentoItem {
  title: string;
  text: string;
}

export interface Herramienta {
  title: string;
  desc: string;
  color: string;
}

export interface Beneficio {
  title: string;
  text: string;
  metric: string;
}